# repair-platform/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个维修平台的示例服务器。",
        "title": "维修平台 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API 支持",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/feedback": {
            "post": {
                "description": "允许用户为特定的维修请求提交反馈",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "反馈"
                ],
                "summary": "提交维修反馈",
                "parameters": [
                    {
                        "description": "反馈内容",
                        "name": "feedback",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Feedback"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "反馈提交成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入数据无效",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "未找到维修请求",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "提交反馈失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/feedback/{id}": {
            "get": {
                "description": "允许管理员或用户查看特定维修请求的反馈",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "反馈"
                ],
                "summary": "查看维修反馈",
                "parameters": [
                    {
                        "type": "string",
                        "description": "维修请求ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "反馈记录",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Feedback"
                            }
                        }
                    },
                    "404": {
                        "description": "未找到此维修请求的反馈",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "检索反馈失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/repair_requests": {
            "get": {
                "description": "管理员查看所有维修请求的列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "维修请求"
                ],
                "summary": "查看维修请求列表",
                "responses": {
                    "200": {
                        "description": "维修请求列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RepairRequest"
                            }
                        }
                    },
                    "500": {
                        "description": "检索维修请求失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "用户提交新的维修请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "维修请求"
                ],
                "summary": "提交维修请求",
                "parameters": [
                    {
                        "description": "维修请求内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RepairRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "维修请求提交成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "输入数据无效",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "提交维修请求失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/repair_requests/{id}": {
            "put": {
                "description": "管理员根据请求ID更新维修请求信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "维修请求"
                ],
                "summary": "更新维修请求",
                "parameters": [
                    {
                        "type": "string",
                        "description": "维修请求ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新的维修请求内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RepairRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "维修请求更新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "输入数据无效",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "未找到维修请求",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "更新维修请求失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Feedback": {
            "type": "object",
            "properties": {
                "comments": {
                    "description": "反馈评论",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "description": "评分",
                    "type": "integer"
                },
                "repairID": {
                    "description": "关联的维修请求ID",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "description": "用户ID",
                    "type": "integer"
                }
            }
        },
...
        "models.RepairRequest": {
            "type": "object"
        },
...
    }
}